requirements.txt
pillow==9.4.0
pilmoji==2.0.2
flask==3.0.0
requests==2.28.2
google-generativeai